requirements.txt
streamlit==1.42.0
google-genai==0.3.0
pydantic==2.10.6
